% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_label.R
\name{add_label}
\alias{add_label}
\title{Add label to base graphic panels}
\usage{
add_label(label, xfrac = 0.01, yfrac = 0.07, pos = 4, ...)
}
\arguments{
\item{label}{string, label to be added to plot}

\item{xfrac}{x-coordinate location, in user coordinates}

\item{yfrac}{y-coordinate location, in user coordinates}

\item{pos}{position specifier for label, see \code{\link[graphics]{text}}}

\item{...}{passed to \code{graphics::text}}
}
\value{
Function only useful for side effect of adding a label.
}
\description{
Adds a label to a base plot using the "user coordinates" of the plotting
 region. The default is to add the label in the upper left corner.
}
\examples{
 plot(1:10)
 add_label("(a) cool label")

}
\seealso{
\code{\link[graphics]{text}}
 \code{\link[graphics]{par}}
}
\author{
Michael Malick
}
