% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/aic_c.R
\name{aic_c}
\alias{aic_c}
\title{Compute AICc}
\usage{
aic_c(model, k = 2)
}
\arguments{
\item{model}{linear model object produced using the \code{lm()} function}

\item{k}{penalty term for additional parameters. k = 2 is standard usage.}
}
\value{
Returns a list:
    \itemize{
        \item{edf: Effective degrees of freedom}
        \item{AIC: AIC value}
        \item{AICc: AICc value}}
}
\description{
This function computes the AIC corrected for small samples sizes (AICc) and
    includes the additive constant. Currently only one object can be analyzed at
    a time.
}
\examples{
## Create some data
x <- rnorm(10)
y <- rnorm(10)

## Fit linear model
fit <- lm(y ~ x)

## Compute AICc
aic_c(fit)
}
\author{
Michael Malick
}
\references{
Burnham, K. P., and D. R. Anderson. 2002. Model selection and multimodel
    inference: a practical information-theoretical approach, 2nd edition.
    Springer-Verlag, New York.
}
\seealso{
\code{\link{AIC}}
}

