% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_normality.R
\name{check_normality}
\alias{check_normality}
\title{Check the normality assumption of a vector}
\usage{
check_normality(x, panel = TRUE, na.rm = FALSE, shapiro.test = FALSE)
}
\arguments{
\item{x}{numeric vector}

\item{panel}{logical, if \code{TRUE} a multi-panel graphic is produced, otherwise each
graphic is printed individually.}

\item{na.rm}{logical, should NA values be removed}

\item{shapiro.test}{logical, should a Shapiro-Wilks test be run}
}
\value{
Prints a series of plots:
 \item{Histogram}{Histogram of data with empirical density overlay (red)
                  and a theoretical normal PDF overlay (grey)}
 \item{QQ plot}{Normal QQ plot}
 \item{CDF}{Empirical (red dots) and theoretical normal (grey line)
            cumulative distributions}

 If \code{shapiro.test}, then the results of the test are returned
}
\description{
Check the normality assumption of univariate data or model residuals using
 graphical methods.
}
\examples{
 check_normality(rnorm(100))
 check_normality(rnorm(100), panel = FALSE)
 check_normality(rlnorm(100))

}
\seealso{
\code{\link[stats]{dnorm}}
 \code{\link[stats]{pnorm}}
 \code{\link[stats]{density}}
 \code{\link[stats]{plot.ecdf}}
 \code{\link[stats]{shapiro.test}}
}
\author{
Michael Malick
}
