% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_outliers.R
\name{check_outliers}
\alias{check_outliers}
\title{Find potential outliers in a numeric vector}
\usage{
check_outliers(x, method = "iqr", range = 3, na.rm = FALSE, plot = TRUE)
}
\arguments{
\item{x}{numeric vector}

\item{method}{method for finding potential outliers}

\item{range}{multiplier of IQR for outlier threshold}

\item{na.rm}{logical, should NA values be removed}

\item{plot}{logical, should a plot of results be produced}
}
\value{
Returns a list with elements:
   \item{outlier}{logical vector the same length as \code{x} where
         \code{TRUE} indicates point is outside threshold values and
         \code{FALSE} otherwise}
   \item{index}{indices of x that are outside threshold}
   \item{threshold}{vector of length two giving lower and upper
         thresholds}
   \item{n.outlier}{number of points outside threshold values}

 If \code{plot = TRUE}, then a plot of \code{x} is produced with potential
 outliers highlighted in red. Dashed horizontal lines indicate outlier
 thresholds.
}
\description{
Identifies potential outliers in a numeric vector based on threshold values
 computed from input data.
}
\details{
Currently only \code{iqr} method is implemented.

 Method details:
     \itemize{
       \item{\code{iqr} identifies an "outlier threshold" based on the
                  interquartile range (IQR) of the data. The upper/lower
                  threshold values are calculated as +/- IQR * range.}
     }
}
\examples{
 check_outliers(c(rnorm(100), 10, -10))
 check_outliers(exp(rnorm(100)), range = 5)
 check_outliers(exp(rnorm(100)))

}
\seealso{
\code{\link[stats]{quantile}}
}
\author{
Michael Malick
}
