% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/error_bar.R
\name{error_bar}
\alias{error_bar}
\title{Error bars for a barplot}
\usage{
error_bar(x, y, ebl, ebu = ebl, length = 0.08, ...)
}
\arguments{
\item{x}{\code{barplot} object}

\item{y}{vector of bar heights}

\item{ebl}{vector of lower error bar limits}

\item{ebu}{vector of upper error bar limits}

\item{length}{length of the horizontal bars}

\item{\dots}{additional parameters controlling the error bar}
}
\value{
Returns a barplot with error bars
}
\description{
Draws error bars on a barplot.
}
\details{
This function currently only works on barplots
}
\examples{
## Bar Heights
x <- c(10,8,13,7)

## Error Bars
up <- c(1,1,1,1)
lo <- c(1,1,1,1)

## Barplot
plt <- barplot(x, ylim = c(0, 15), las = 1)
error_bar(plt, x, up, lo)
}
\author{
Michael Malick
}
\seealso{
\code{\link{barplot}}
}

