% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/jet_colors.R
\name{jet_colors}
\alias{jet_colors}
\title{Matlab jet color palette}
\usage{
jet_colors(n)
}
\arguments{
\item{n}{number of levels the palette should be broken into. For
    example n = 100 creates 100 shades ranging over the specified
    colors.}
}
\value{
This function outputs a vector of colors with length equal to n.
}
\description{
Replica of the `jet colors' palette in matlab.
    Produces a color palette using the function colorRampPalette that
    starts at darkblue moves through cyan, green, yellow, and ends at
    darkred. This color palette works well for representing mapped
    oceanographic data such as SST or chl-a.
}
\examples{
## Create a matrix of data
mat <- matrix(data = seq(0.1, 100, .1), nrow = 100, ncol = 10)

## Plot data using jet_colors()
image(mat, col = jet_colors(100))
}
\author{
Michael Malick
}
\seealso{
\code{\link{colorRampPalette}}
    \code{\link{heat.colors}}
    \code{\link{topo.colors}}
    \code{\link{terrain.colors}}
}

