% Generated by roxygen2 (4.0.1): do not edit by hand
\name{mm.par}
\alias{mm.par}
\title{Par settings for lattice graphics}
\usage{
mm.par(...)
}
\arguments{
\item{\dots}{A list of parameters to pass to \code{par.settings}. This
    allows the user to change additional parameters not set by
    \code{mm.par} or to change the settings in \code{mm.par}.}
}
\value{
Returns a list of visually pleasing par setting for lattice graphics.
}
\description{
Returns a list of par settings for lattice that can be passed
    to the \code{par.settings} argument.
}
\details{
A list of parameters supplied to the function should be in the
    same format as for \code{par.settings}.
}
\examples{
\dontrun{
require(lattice)

xyplot(Petal.Width ~ Petal.Length | Species, data = iris,
    par.settings = mm.par())

xyplot(Petal.Width ~ Petal.Length | Species, data = iris,
    par.settings = mm.par(), grid = TRUE)

xyplot(Petal.Width ~ Petal.Length | Species, data = iris,
    par.settings = mm.par(list(fontsize = list(text = 16),
    par.main.text = list(col = "grey50"),
    plot.symbol = list(pch="l"))),
    main = "Iris")

barchart(Species ~ Petal.Length, data = iris,
    par.settings = mm.par())

barchart( ~ Petal.Length, data = iris, groups = Species,
    par.settings = mm.par(), auto.key = TRUE,
    panel = function(x, y, ...) {
        panel.grid(-1, -1)
        panel.barchart(x, y, ...)
    })

dotplot(Species ~ Petal.Length, data = iris,
    par.settings = mm.par())
}
}
\author{
Michael Malick
}
\seealso{
\code{\link{trellis.par.get}}
    \code{\link{trellis.par.set}}
    \code{\link{lattice.options}}
    \code{\link{xyplot}}
}

