% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/norm_diag.R
\name{norm_diag}
\alias{norm_diag}
\title{Explore the normality assumption of a vector}
\usage{
norm_diag(x, shapiro.test = FALSE, na.rm = TRUE, main.title = "")
}
\arguments{
\item{x}{vector of numeric data}

\item{shapiro.test}{logical, should a Shaprio-Wilks test be run}

\item{na.rm}{logical, should NA values be removed before plotting}

\item{main.title}{main title of plot}
}
\value{
Returns a plotting device with four diagnostic plots:
    \itemize{
        \item{histogram: includes an empirical density estimate}
        \item{boxplot}
        \item{normal qq plot}
        \item{empirical cdf: includes the corresponding normal
            cumulative density function}}
}
\description{
Examine the distribution of data and check against normal distribution.
}
\examples{
x <- rnorm(100)
norm_diag(x)
}
\author{
Michael Malick
}
\seealso{
\code{\link{shapiro.test}}
    \code{\link{plot.ecdf}}
    \code{\link{boxplot}}
    \code{\link{hist}}
    \code{\link{density}}
}

