% Generated by roxygen2 (4.0.2): do not edit by hand
\name{panel.mean}
\alias{panel.mean}
\title{Add factor means to a lattice plot}
\usage{
panel.mean(x, y, fun = mean, horizontal = TRUE, lwd = reference.line$lwd,
  lty = reference.line$lty, col, col.line = reference.line$col,
  type = "l", ...)
}
\arguments{
\item{x}{x values}

\item{y}{y values}

\item{fun}{\code{mean} function is used for each panel}

\item{horizontal}{logical}

\item{lwd}{line width}

\item{lty}{line type}

\item{col}{color}

\item{col.line}{line color}

\item{type}{points or line}

\item{\dots}{other plotting parameters}
}
\value{
Used as part of an \code{\link{xyplot}} call
}
\description{
Panel function for lattice plots to add a mean bar onto an xyplot
    that plots a numeric ~ factor, where the the mean represents the
    mean of the factor.
}
\examples{
require(lattice)

xyplot(Sepal.Length ~ Species, data = iris,
    panel = function(x, y, col, ...) {
	       panel.grid(h=4, v=0, col = "grey80", ...)
		   panel.xyplot(x, y, col = "grey40", ...)
		   panel.mean(x, y, type = "p", horizontal = F, col = 1,
		       pch = "-", cex = 10, ...)
    })
}
\author{
Michael Malick
}
\seealso{
\code{\link{lattice}}
    \code{\link{xyplot}}
    \code{\link{panel.abline}}
}

