% Generated by roxygen2 (4.0.2): do not edit by hand
\name{theme_white}
\alias{theme_white}
\title{ggplot2 White Theme}
\usage{
theme_white(base_size = 12, base_family = "", axis_text_size = 8)
}
\arguments{
\item{base_size}{base text size}

\item{base_family}{base text family}

\item{axis_text_size}{size of axis tick labels}
}
\value{
Part of a ggplot2 object
}
\description{
A ggolot2 theme that is all white with grey grid lines
}
\examples{
\dontrun{
library(ggplot2)

ggplot(mpg, aes(cty, hwy)) +
    geom_point(colour = "red") +
    theme_white()

ggplot(mpg, aes(cty, hwy)) +
    geom_point() +
    facet_wrap(~class) +
    theme_white()

ggplot(mpg, aes(cty, hwy)) +
    geom_point() +
    facet_grid(year~class) +
    theme_white()'
}
}
\author{
Michael Malick
}
\seealso{
\code{\link{ggplot} \link{theme_gray}}
}

