% Generated by roxygen2 (4.0.1): do not edit by hand
\name{boot.median}
\alias{boot.median}
\title{Non-parametric bootstrap of the median}
\usage{
boot.median(x, rep = 999)
}
\arguments{
\item{x}{a numeric vector}

\item{rep}{number of iterations}
}
\value{
Returns a list:
    \itemize{
        \item{boot.mean}{mean of the bootstrap samples}
        \item{boot.sd}{standard deviation of the bootstrap samples}
        \item{bias}{the mean of the bootstrap samples minus the
            median of the original data}
        \item{distribution}{vector of the bootstrap samples}}
}
\description{
This function performs a non-parametric bootstrap of a vector with
    replacement.
}
\examples{
x <- 1:10
boots <- boot.median(x)

hist(boots$distribution, col = "grey20")
abline(v = boots$boot.mean, lwd = 2, col = 2)
}
\author{
Michael Malick
}
\seealso{
\code{\link{median}} \code{\link{sample}}
}

