% Generated by roxygen2 (4.0.2): do not edit by hand
\name{conf.int}
\alias{conf.int}
\title{Compute confidence intervals}
\usage{
conf.int(x, alpha = 0.975, na.rm = FALSE)
}
\arguments{
\item{x}{numeric vector}

\item{alpha}{significance level}

\item{na.rm}{logical, should NA's be removed. Function fails if NA's are
    present and na.rm = FALSE.}
}
\value{
Returns a list:
    \itemize{
        \item{Mean: mean of the vector}
        \item{SD: standard deviation of the vector}
        \item{CI: confidence interval}
        \item{Upper: upper confidence bound}
        \item{Lower: lower confidence bound}}
}
\description{
Function to calculate confidence intervals (two-tailed) based on
    the t-distribution.
}
\examples{
x <- 1:10
conf.int(x)

y <- c(1, 2, 3, NA, 4, 5:10)
conf.int(y, na.rm = TRUE)
}
\author{
Michael Malick
}
\seealso{
\code{\link{gt}}
}

