% Generated by roxygen2 (4.0.2): do not edit by hand
\name{cor.prob}
\alias{cor.prob}
\title{Pairwise correlation matrix with p-values}
\usage{
cor.prob(x)
}
\arguments{
\item{x}{Matrix or dataframe}
}
\value{
Returns a matrix
}
\description{
The function produces a matrix of all pairwise correlations with
    the correlations below the main diagonal and the p-values from
    pairwise deletion significant tests above the main diagonal.
}
\details{
Correlations are performed among the columns of the matrix or dataframe.
}
\examples{
x <- rnorm(25)
mat <- matrix(data = x, nrow = 5, ncol = 5, dimnames = list(
    c("A", "B","C", "D", "E"), c("M", "N","O", "P", "Q")))

cor.prob(mat)

# Iterative version of cor Function
cor(mat, use = "pairwise.complete.obs")
}
\author{
Michael Malick
}
\seealso{
\code{\link{cor}}
}

