% Generated by roxygen2 (4.0.1): do not edit by hand
\name{cross.validate}
\alias{cross.validate}
\title{Cross validation routine for glm's}
\usage{
cross.validate(fit, observed, data, k = 1)
}
\arguments{
\item{fit}{fitted glm object}

\item{observed}{observed x values, need specify columns e.e., dat[1]}

\item{data}{data used to fit glm model}

\item{k}{number of data points to leave out}
}
\value{
Returns a list:
    \itemize{
        \item{MSE: mean squared prediction error}
        \item{comp1: prediction standard deviation}}
}
\description{
Function to compute prediction standard errors for general linear
    models using leave-one-out cross validation techniques
}
\details{
Modified code from cv.glm in boot package
}
\examples{
set.seed(29)
x <- rnorm(25)
y <- rnorm(25)
dat <- data.frame(x = x, y = y)

fit <- glm(y ~ x) # same result as lm()
cross.validate(fit = fit, observed = dat[1], data = dat)
}
\author{
Michael Malick
}
\seealso{
\code{\link{glm}} \code{\link{predict}}
}

