% Generated by roxygen2 (4.0.2): do not edit by hand
\name{define.gridcells}
\alias{define.gridcells}
\title{Define gridcells within a study region}
\usage{
define.gridcells(sub.cells = NULL, cell.size = 2, lat.min = -90,
  lat.max = 90, lon.min = -180, lon.max = 180, plot = TRUE,
  locations = NULL)
}
\arguments{
\item{sub.cells}{optional dataframe giving the indices of the grid cells to subset within the
study region. It's expected that that the grid coordinates have 1,1 in the
top lefthand corner. If NULL all grid cells are used}

\item{cell.size}{grid cell size in degrees}

\item{lat.min}{minimum latitude of study region}

\item{lat.max}{maximum latitude of study region}

\item{lon.min}{minimum longitude of study region}

\item{lon.max}{maximum longitude of study region}

\item{plot}{logical, should a plot of the grid cells be created}

\item{locations}{optional coordinates to plot on map}
}
\value{
The output of the function is a dataframe that gives the unique
    row and column identifiers for each grid cell, the unique grid
    cell id for all grid cells (matrix.id) and the subsetted grid
    cells (cell.id), the center coordinates of the grid cell (lat and
    lon), and the max and min coordinates for each grid cell.
}
\description{
This function creates a grid over a defined study region and
    returns the coordinates for the center of the gridcells. It can
    also subset these gridcells according to predefined grid cell
    coordinates.
}
\details{
This functions takes as input the boundaries of the study region
    and the size of the desired grid cells in degrees. The function
    creates a grid of cells over the entire region giving each grid
    cell a unique id (matrix.id). The function treats this grid like
    a matrix with each cell having a unique column and row
    identifier. The sub.cells argument takes as input a dataframe
    that gives the row and column indices of specific grid cells to
    subset. The first index (1,1) should be in the upper
    lefthand corner and filling by row moving to the right. The last
    cell should be in the lower righthand corner. If sub.cells is
    supplied only data for the given cells are returned. If sub.cells
    is NULL data for all cells in the region are returned.

    The function expects the study region to be defined within the
    coordinates: Lat: -90:90 Lon: -180:180.

    Requires the reshape2 and plyr packages.
}
\examples{
define.gridcells(plot = FALSE)
}
\author{
Michael Malick
}

