% Generated by roxygen2 (4.0.2): do not edit by hand
\name{lm.diag}
\alias{lm.diag}
\title{Linear model diagnostics}
\usage{
lm.diag(x, main.title = "Residual Diagnostic Plots")
}
\arguments{
\item{x}{linear model object}

\item{main.title}{title for plotting device}
}
\value{
Returns a plotting device with eight standard diagnostic plots for linear
    models:
    \itemize{
        \item{histogram of the residuals}
        \item{normal qq plot of residuals}
        \item{residuals vs fitted values}
        \item{standardized residuals vs fitted values}
        \item{studentized residuals vs fitted values}
        \item{autocorrelation function}
        \item{cooks distance vs fitted}
        \item{leverages vs fitted}}
}
\description{
This function produces eight standard diagnostic plots used to
    assess the assumptions of a linear regression. The plots analyze
    the residuals of the linear model. See also
    \code{link{norm.diag}} for assessing the normality of the
    residuals.
}
\details{
\itemize{
        \item{Standardized and studentized residuals are calculated according
            to the functions in the MASS library}
        \item{Cooks distances give the influence of a point on all fitted values
            (4/(n-p-1) = potentially influential)}
    \item{Leverages give influence of a point on
        the model fit. Leverages over 2p/n are considered influential.
        ACF values +/- 2/sqrt(n) are condidered significant}}
}
\examples{
## Create some data
x <- rnorm(100)
y <- 1:100

## Fit linear model
fitted <- lm(x ~ y)

## Run diagnostics
lm.diag(fitted)
}
\author{
Michael Malick
}
\references{
Kutner, M. H., C. J. Nachtsheim, J. Neter, and W. Li. 2005.
    Applied linear statistical models, 5th edition.
    McGraw-Hill/Irwin, New York.
}
\seealso{
\code{\link{lm}}
    \code{\link{cooks.distance}}
    \code{\link{stdres}}
    \code{\link{studres}}
    \code{\link{norm.diag}}
    \code{\link{influence}}
}

