% Generated by roxygen2 (4.0.2): do not edit by hand
\name{load.ncdf}
\alias{load.ncdf}
\title{Load netCDF file}
\usage{
load.ncdf(datafile, var, lat.name = "lat", lon.name = "lon",
  time.name = "time")
}
\arguments{
\item{datafile}{netCDF file to load}

\item{var}{name of the variable in the netCDF file}

\item{lat.name}{name of the latitude variable in the netCDF file}

\item{lon.name}{name of the longitude variable in the netCDF file}

\item{time.name}{name of the time variable in the netCDF file}
}
\value{
Returns a list:
    \itemize{
        \item{datLat: latitude vector}
        \item{datLon: longitude vector}
        \item{datTime: time vector}
        \item{dim: dimensions of dat}
        \item{dat: data arrays}}
}
\description{
This function loads a netCDF file
}
\examples{
\dontrun{load.ncdf("./data/sst.nc", var = "sst")}
}
\author{
Michael Malick
}

