% Generated by roxygen2 (4.0.2): do not edit by hand
\name{my.ls}
\alias{my.ls}
\title{List objects by class and size}
\usage{
my.ls(pos = 1)
}
\arguments{
\item{pos}{position of the search}
}
\description{
This function is a fancy version of ls() that includes the
    objects class and the objects size.
}
\examples{
x <- 10
y <- iris
my.ls()
}
\author{
Michael Malick
}
\seealso{
\code{\link{ls}}
}

