% Generated by roxygen2 (4.0.1): do not edit by hand
\name{outlier}
\alias{outlier}
\title{Identify outliers in a series of residuals}
\usage{
outlier(x, type = "outlier", p = 0.05, plot.it = TRUE)
}
\arguments{
\item{x}{vector of residuals}

\item{type}{One of \code{"outlier"}, \code{"max.abs"}, or \code{"max.min"}.}

\item{p}{p-value}

\item{plot.it}{logical, should the results be plotted}
}
\value{
Returns a list with two components:
    \itemize{
        \item{positive}{positive outliers}
        \item{negative}{negative outliers}}

    There are three types:
    \itemize{
        \item{outlier}{
            returns the positions of all residuals that are
            larger than (smaller than) the (1-p/2)th (p/2th) percentile
            of the distribution of the residuals assuming they follow a
            normal distribution.
        }

        \item{max.abs}{
            returns the position and sign of the maximum absolute
            residual
        }

        \item{max.min}{
            returns the positions of the maximum and minimum residual
        }
    }
}
\description{
Identifies outliers in a series of residuals according their index number in
    the vector.
}
\examples{
set.seed(29)
    x <- rnorm(25)
    y <- rnorm(25)
    fit <- lm(y ~ x)
    outlier(resid(fit))
}
\author{
Michael Malick
}
\seealso{
\code{\link{qqnorm}}
}

