% Generated by roxygen2 (4.0.1): do not edit by hand
\name{rand.mean}
\alias{rand.mean}
\title{Randomization test for two means}
\usage{
rand.mean(x, y, R = 1000, one.sided = FALSE)
}
\arguments{
\item{x}{numeric vector, sample 1}

\item{y}{numeric vector, sample 2}

\item{R}{number of iterations}

\item{one.sided}{logical, should the test be one-tailed}
}
\value{
Returns a histogram of the null distribution indicating the
    observed difference with a red vertical line and a list with
    components:
    \itemize{
        \item{null.distr: null distribution}
        \item{data: a list giving the two vectors}
        \item{T: absolute value of the differences between the
            means of the samples}
        \item{p.value: the p-value of the test}
        \item{test: tells if its a one or two sided test}
    }
}
\description{
Function to test the difference between means of two samples
    using randomization.
}
\examples{
x <- 1:100
y <- 11:110
rand.mean(x, y)
}
\author{
Michael Malick
}
\seealso{
\code{\link{mean}}
    \code{\link{sample}}
}

